SphericalHarmonicArrays.jl

SphericalHarmonicArrays.SHArray — Type

SHArray(arr::AbstractArray{T,N}, modes::Tuple, shdims::NTuple{NSH, Int}) where {T,N,NSH}

Create a wrapper around an array such that the dimensions specified by shdims may be indexed using a Tuple representing a collection of spherical harmonic degrees. Often this would be a spherical harmonic mode represented by (l,m).

Use the iterators provided by SphericalHarmonicModes in modes to generate the map between the indices of the array and Tuples of spherical harmonic degrees.

Examples

julia> s = SHArray(reshape(1:4, 2, 2), (LM(0:1,0:0), 2))
2×2 SHArray{Int64,2,Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}},Tuple{LM,Base.OneTo{Int64}},1}:
 1  3
 2  4

julia> s.modes[1] |> collect
2-element Array{Tuple{Int64,Int64},1}:
 (0, 0)
 (1, 0)

julia> s[(1,0), 1]
2

See also: SHVector, SHMatrix

SphericalHarmonicArrays.SHArray — Method

SHArray{T}(modes::Tuple, [shdims::Tuple])

Return an SHArray wrapper around a parent array of the appropriate size with elements of type T, such that the dimensions specified by shdims may be indexed using Tuples of spherical harmonic degrees. The elements of the parent array are set to zero by default.

SHArray(modes::Tuple, [shdims::Tuple])

Return an SHArray of the appropriate size and element type ComplexF64.

Examples

julia> s = SHArray((LM(0:1,0:0), 2))
2×2 SHArray{Complex{Float64},2,Array{Complex{Float64},2},Tuple{LM,Base.OneTo{Int64}},1}:
 0.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im

SphericalHarmonicArrays.SHMatrix — Method

SHMatrix(arr::AbstractMatrix, modes::NTuple{2,ModeRange})
SHMatrix(arr::AbstractMatrix, ax1::ModeRange, ax2::ModeRange)

Return a wrapper that maps the indices of the parent matrix along each axis to a Tuple of spherical harmonic degrees. Often these Tuples would be pairs of spherical harmonic modes (l,m).

Examples

julia> s = SHMatrix(reshape(1:4, 2,2), (LM(0:1, 0:0), LM(1:2,0:0)))
2×2 SHArray{Int64,2,Base.ReshapedArray{Int64,2,UnitRange{Int64},Tuple{}},Tuple{LM,LM},2}:
 1  3
 2  4

julia> s[(1,0),(2,0)]
4

See also: SHVector, SHArray

SphericalHarmonicArrays.SHVector — Method

SHVector(arr::AbstractVector, mode::ModeRange)

Return a wrapper that maps the indices of the parent vector to a Tuple of spherical harmonic degrees. Often the Tuple would be a pair of spherical harmonic modes (l,m).

Examples

julia> s = SHVector(1:3, LM(1:1))
3-element SHArray{Int64,1,UnitRange{Int64},Tuple{LM},1}:
 1
 2
 3

julia> s.modes[1] |> collect
3-element Array{Tuple{Int64,Int64},1}:
 (1, -1)
 (1, 0)
 (1, 1)

julia> s[(1,0)]
2

See also: SHMatrix, SHArray